Assignement 1 complete

+++ src/main/java/com/eventhub/eventhubapi/service/impl/EventServiceImpl.java
@@ -1,47 +1,29 @@
 package com.eventhub.eventhubapi.service.impl;
 
+import com.eventhub.eventhubapi.dto.CreateEventDTO;
 import com.eventhub.eventhubapi.dto.EventDTO;
+import com.eventhub.eventhubapi.exception.EventNotFoundException;
 import com.eventhub.eventhubapi.model.Event;
 import com.eventhub.eventhubapi.repository.EventRepository;
 import com.eventhub.eventhubapi.service.EventService;
 import org.springframework.stereotype.Service;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 
 import java.time.LocalDateTime;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-/**
- * EventServiceImpl
- *
- * Service layer implementation responsible for
- * handling business logic related to Event operations.
- *
- * This layer acts as a bridge between the Controller
- * and Repository layers.
- */
 @Service
 public class EventServiceImpl implements EventService {
 
-    /**
-     * Repository dependency used for data persistence.
-     */
     private final EventRepository repository;
 
-    /**
-     * Constructor-based dependency injection.
-     *
-     * @param repository the event repository implementation
-     */
     public EventServiceImpl(EventRepository repository) {
         this.repository = repository;
     }
 
-    /**
-     * Converts Event entity to EventDTO.
-     *
-     * @param event the Event entity
-     * @return corresponding EventDTO
-     */
     private EventDTO toDTO(Event event) {
         return new EventDTO(
                 event.getId(),
@@ -54,16 +36,7 @@ public class EventServiceImpl implements EventService {
         );
     }
 
-    /**
-     * Converts EventDTO to Event entity.
-     *
-     * ID is set to null so that the repository
-     * generates a new one if necessary.
-     *
-     * @param dto the EventDTO
-     * @return corresponding Event entity
-     */
-    private Event toEntity(EventDTO dto) {
+    private Event toEntity(CreateEventDTO dto) {
         return new Event(
                 null,
                 dto.getName(),
@@ -77,54 +50,127 @@ public class EventServiceImpl implements EventService {
         );
     }
 
-    /**
-     * Retrieves all events.
-     *
-     * @return list of EventDTO objects
-     */
     @Override
-    public List<EventDTO> getAllEvents() {
-        return repository.findAll()
+    @Cacheable("events")
+    public List<EventDTO> getAllEvents(
+            int page,
+            int size,
+            String sort,
+            String category,
+            Double minPrice,
+            Double maxPrice,
+            String startDate,
+            String endDate
+    ) {
+
+        List<Event> events = repository.findAll();
+
+        //Filter by category
+        if (category != null) {
+            events = events.stream()
+                    .filter(e -> e.getCategory().equalsIgnoreCase(category))
+                    .toList();
+        }
+
+        //Filter by minimum price
+        if (minPrice != null) {
+            events = events.stream()
+                    .filter(e -> e.getTicketPrice().doubleValue() >= minPrice)
+                    .toList();
+        }
+
+        //Filter by maximum price
+        if (maxPrice != null) {
+            events =  events.stream()
+                    .filter(e -> e.getTicketPrice().doubleValue() <= maxPrice)
+                    .toList();
+        }
+
+
+        // Filter by start date
+        if (startDate != null) {
+            LocalDateTime start = LocalDateTime.parse(startDate);
+            events = events.stream()
+                    .filter(e -> e.getEventDate().isAfter(start))
+                    .toList();
+        }
+
+        // Filter by end date
+        if (endDate != null) {
+            LocalDateTime start = LocalDateTime.parse(startDate);
+            events = events.stream()
+                    .filter(e -> e.getEventDate().isAfter(start))
+                    .toList();
+        }
+
+        //Pagination
+        int startIndex = page * size;
+        int endIndex = Math.min(startIndex + size, events.size());
+
+        if (startIndex > events.size()) {
+            return List.of();
+        }
+
+        // Sorting
+        String[] sortParts = sort.split(",");
+        String sortField = sortParts[0];
+        String sortDirection = sortParts.length > 1 ? sortParts[1] : "asc";
+
+        Comparator<Event> comparator;
+
+        switch (sortField) {
+            case "ticketPrice":
+                comparator = Comparator.comparing(Event::getTicketPrice);
+                break;
+            case "eventDate":
+                comparator = Comparator.comparing(Event::getEventDate);
+                break;
+            case "category":
+                comparator = Comparator.comparing(Event::getCategory, String.CASE_INSENSITIVE_ORDER);
+                break;
+            case "name":
+            default:
+                comparator = Comparator.comparing(Event::getName, String.CASE_INSENSITIVE_ORDER);
+                break;
+        }
+
+        if ("desc".equalsIgnoreCase(sortDirection)) {
+            comparator = comparator.reversed();
+        }
+
+        events = events.stream()
+                .sorted(comparator)
+                .toList();
+
+        return events.subList(startIndex, endIndex)
                 .stream()
                 .map(this::toDTO)
-                .collect(Collectors.toList());
+                .toList();
     }
 
-    /**
-     * Retrieves an event by ID.
-     *
-     * @param id event identifier
-     * @return EventDTO if found, otherwise null
-     */
     @Override
     public EventDTO getEventById(Long id) {
         Event event = repository.findById(id);
-        return event != null ? toDTO(event) : null;
+        if (event == null) {
+            throw new EventNotFoundException("Event not found with id: " + id);
+        }
+        return toDTO(event);
     }
 
-    /**
-     * Creates a new event.
-     *
-     * @param dto event data received from client
-     * @return created EventDTO
-     */
     @Override
-    public EventDTO createEvent(EventDTO dto) {
+    @CacheEvict(value = "events", allEntries = true)
+    public EventDTO createEvent(CreateEventDTO dto) {
         Event event = toEntity(dto);
         return toDTO(repository.save(event));
     }
 
-    /**
-     * Updates an existing event.
-     *
-     * @param id  event identifier
-     * @param dto updated event data
-     * @return updated EventDTO or null if not found
-     */
     @Override
-    public EventDTO updateEvent(Long id, EventDTO dto) {
+    @CacheEvict(value = "events", allEntries = true)
+    public EventDTO updateEvent(Long id, CreateEventDTO dto) {
         Event existing = repository.findById(id);
-        if (existing == null) return null;
+        if (existing == null) {
+            throw new EventNotFoundException("Event not found with id: " + id);
+        }
 
         existing.setName(dto.getName());
         existing.setDescription(dto.getDescription());
@@ -137,13 +183,13 @@ public class EventServiceImpl implements EventService {
         return toDTO(repository.save(existing));
     }
 
-    /**
-     * Deletes an event by ID.
-     *
-     * @param id event identifier
-     */
     @Override
+    @CacheEvict(value = "events", allEntries = true)
     public void deleteEvent(Long id) {
+        Event existing = repository.findById(id);
+        if (existing == null) {
+            throw new EventNotFoundException("Event not found with id: " + id);
+        }
         repository.delete(id);
     }
 }
